Update memcachedbetween dependency
Use ServerWrapper and ConnectionWrapper

Refactor message passing logic to messenger struct
