+++ proxy/proxy_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"github.com/DataDog/datadog-go/statsd"
 	"github.com/coinbase/redisbetween/config"
-	"github.com/coinbase/redisbetween/handlers"
+	"github.com/coinbase/redisbetween/messenger"
 	"github.com/go-redis/redis/v8"
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
@@ -89,7 +89,7 @@ func TestPipelinedCommands(t *testing.T) {
 		go func(index int, t *testing.T) {
 			var j int
 			ind := strconv.Itoa(index)
-			commands := []command{{cmd: "get", args: []string{string(handlers.PipelineSignalStartKey)}, res: "get 🔜: redis: nil"}}
+			commands := []command{{cmd: "get", args: []string{string(messenger.PipelineSignalStartKey)}, res: "get 🔜: redis: nil"}}
 			for {
 				j++
 				if j == 20 {
@@ -99,7 +99,7 @@ func TestPipelinedCommands(t *testing.T) {
 				commands = append(commands, command{cmd: "set", args: []string{s, "hi"}, res: "set " + s + " hi: OK"})
 				commands = append(commands, command{cmd: "get", args: []string{s}, res: "get " + s + ": hi"})
 			}
-			commands = append(commands, command{cmd: "get", args: []string{string(handlers.PipelineSignalEndKey)}, res: "get 🔚: redis: nil"})
+			commands = append(commands, command{cmd: "get", args: []string{string(messenger.PipelineSignalEndKey)}, res: "get 🔚: redis: nil"})
 			assertResponsePipelined(t, commands, client)
 			wg.Done()
 		}(i, t)
